+++ src/aria/simple-combobox/public-api.ts
@@ -1,0 +1,18 @@
+/**
+ * @license
+ * Copyright Google LLC All Rights Reserved.
+ *
+ * Use of this source code is governed by an MIT-style license that can be
+ * found in the LICENSE file at https://angular.dev/license
+ */
+
+export {Combobox} from './simple-combobox';
+export {ComboboxPopup} from './simple-combobox-popup';
+export {ComboboxWidget} from './simple-combobox-widget';
+
+// This needs to be re-exported, because it's used by the combobox components.
+// See: https://github.com/angular/components/issues/30663.
+export {
+  DeferredContent as ɵɵDeferredContent,
+  DeferredContentAware as ɵɵDeferredContentAware,
+} from '../private';
